requirements.txt
asgiref==3.9.2
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
numpy==2.3.3
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
python-dateutil==2.8.2
django-cors-headers==4.3.1
